requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
deep-translator
torch
transformers
networkx
jieba
opencc-python-reimplemented
